refactor(fe): :zap: get categories and regions list from server side
