+++ apps/fe/src/components/layout/header/index.tsx
@@ -2,21 +2,19 @@
 
 import './header.css';
 
-import React, { useState, ReactNode, useEffect, useRef } from 'react';
-import { Layout, Menu, Input, Button, Drawer, Grid, Dropdown, Avatar, Tooltip } from 'antd';
-import { SearchOutlined, UserOutlined, MenuOutlined, DownOutlined } from '@ant-design/icons';
+import React, { useState, ReactNode } from 'react';
+import { Layout, Menu, Button, Drawer, Grid } from 'antd';
+import { MenuOutlined } from '@ant-design/icons';
 import Link from 'next/link';
 import Image from 'next/image';
-import { usePathname, useRouter } from 'next/navigation';
-import { stringifyTableParams, useGetIdentity } from '@refinedev/core';
 import { ItemType, MenuItemType } from 'antd/lib/menu/interface';
-import { signOut } from 'next-auth/react';
 
 import { MovieTypeEnum } from 'apps/api/src/app/movies/movie.constant';
-import type { UserDto } from 'apps/api/src/app/users/dtos/user.dto';
 
 import { RouteNameEnum } from '@/constants/route.constant';
-import { getOptimizedImageUrl } from '@/libs/utils/movie.util';
+import { stringifyTableParams } from '@/libs/utils/url.util';
+import HeaderSearch from './header-search';
+import HeaderUser from './header-user';
 
 const { Header } = Layout;
 const { useBreakpoint } = Grid;
@@ -124,28 +122,8 @@ export type HeaderProps = {
 };
 
 export default function HeaderCom({ categoryMenu = [], regionMenu = [] }: HeaderProps) {
-    const pathname = usePathname();
     const [drawerVisible, setDrawerVisible] = useState(false);
-    const [searchVisible, setSearchVisible] = useState(false);
-    const [searchFocused, setSearchFocused] = useState(false);
     const screens = useBreakpoint();
-    const [searchValue, setSearchValue] = useState('');
-    const router = useRouter();
-    const { data: user, isLoading } = useGetIdentity<UserDto>();
-    const searchRef = useRef<HTMLDivElement>(null);
-
-    useEffect(() => {
-        const handleClickOutside = (event: MouseEvent) => {
-            if (searchRef.current && !searchRef.current.contains(event.target as Node)) {
-                setSearchFocused(false);
-            }
-        };
-
-        document.addEventListener('mousedown', handleClickOutside);
-        return () => {
-            document.removeEventListener('mousedown', handleClickOutside);
-        };
-    }, []);
 
     const navItems = [
         ...baseNavItems,
@@ -167,33 +145,6 @@ export default function HeaderCom({ categoryMenu = [], regionMenu = [] }: Header
 
     const onCloseDrawer = () => {
         setDrawerVisible(false);
-    };
-
-    const toggleSearch = () => {
-        setSearchVisible(!searchVisible);
-    };
-
-    const handleSearchIconClick = () => {
-        const searchQuery = stringifyTableParams({
-            filters: [
-                {
-                    field: 'keywords',
-                    operator: 'eq',
-                    value: searchValue,
-                },
-            ],
-            sorters: [],
-        });
-
-        router.push(`${RouteNameEnum.MOVIE_LIST_PAGE}?${searchQuery}`);
-
-        if (!screens.md) {
-            toggleSearch();
-        }
-    };
-
-    const handleSearch = () => {
-        handleSearchIconClick();
     };
 
     return (
@@ -241,44 +192,7 @@ export default function HeaderCom({ categoryMenu = [], regionMenu = [] }: Header
                 />
             )}
             <div style={{ display: 'flex', alignItems: 'center', position: 'relative' }}>
-                <>
-                    {screens.md ? (
-                        <div ref={searchRef} style={{ position: 'relative' }}>
-                            <Input
-                                placeholder="Tìm kiếm"
-                                suffix={
-                                    <SearchOutlined
-                                        onClick={handleSearchIconClick}
-                                        style={{ cursor: 'pointer' }}
-                                    />
-                                }
-                                style={{
-                                    width: searchFocused ? '18.75rem' : '12.5rem',
-                                    marginRight: '1rem',
-                                    transition: 'all 0.3s ease',
-                                }}
-                                value={searchValue}
-                                onChange={(e) => setSearchValue(e.target.value)}
-                                onFocus={() => setSearchFocused(true)}
-                                onBlur={() => !searchValue && setSearchFocused(false)}
-                                onPressEnter={handleSearch}
-                                allowClear
-                                onClear={() => {
-                                    setSearchValue('');
-                                    setSearchFocused(false);
-                                }}
-                            />
-                        </div>
-                    ) : (
-                        <Button
-                            type="text"
-                            icon={<SearchOutlined />}
-                            onClick={toggleSearch}
-                            style={{ marginRight: '0.5rem' }}
-                        />
-                    )}
-                </>
-
+                <HeaderSearch />
                 {!screens.md && (
                     <Button
                         type="text"
@@ -287,74 +201,7 @@ export default function HeaderCom({ categoryMenu = [], regionMenu = [] }: Header
                         style={{ marginRight: '0.5rem' }}
                     />
                 )}
-                {user && !isLoading ? (
-                    <Dropdown
-                        menu={{
-                            items: [
-                                {
-                                    key: '1',
-                                    label: (
-                                        <Link href={'/nguoi-dung/cap-nhat-thong-tin'}>
-                                            Thông tin tài khoản
-                                        </Link>
-                                    ),
-                                },
-                                {
-                                    key: '2',
-                                    label: <Link href={'/tu-phim'}>Tủ phim</Link>,
-                                },
-                                {
-                                    key: '3',
-                                    label: (
-                                        <Link href={'#'} onClick={() => signOut()}>
-                                            Đăng xuất
-                                        </Link>
-                                    ),
-                                },
-                            ],
-                        }}
-                        trigger={['click']}
-                    >
-                        <Button
-                            type="text"
-                            href={'#'}
-                            onClick={(e) => e.preventDefault()}
-                            style={{ display: 'flex', alignItems: 'center' }}
-                        >
-                            <Avatar
-                                src={
-                                    user.avatar?.url
-                                        ? getOptimizedImageUrl(user.avatar?.url, {
-                                              width: 32,
-                                              height: 32,
-                                              quality: 30,
-                                          })
-                                        : null
-                                }
-                                icon={<UserOutlined />}
-                                style={{ marginRight: '0.3rem' }}
-                            />
-                            {screens?.md && (
-                                <span style={{ marginRight: '0.3rem' }}>{user.fullName}</span>
-                            )}
-                            <DownOutlined />
-                        </Button>
-                    </Dropdown>
-                ) : (
-                    <Tooltip title="Đăng nhập/Đăng kí">
-                        <Button
-                            type="default"
-                            icon={<UserOutlined />}
-                            onClick={() =>
-                                router.push(
-                                    `${RouteNameEnum.LOGIN_PAGE}?to=${encodeURIComponent(
-                                        pathname,
-                                    )}`,
-                                )
-                            }
-                        ></Button>
-                    </Tooltip>
-                )}
+                <HeaderUser />
             </div>
             <Drawer
                 title="Menu"
@@ -367,22 +214,7 @@ export default function HeaderCom({ categoryMenu = [], regionMenu = [] }: Header
                 width={'60%'}
             >
                 <div style={{ padding: '1rem' }}>
-                    <Input
-                        placeholder="Tìm kiếm"
-                        suffix={
-                            <SearchOutlined
-                                onClick={handleSearchIconClick}
-                                style={{ cursor: 'pointer' }}
-                            />
-                        }
-                        style={{ width: '100%', marginBottom: '1rem' }}
-                        value={searchValue}
-                        onChange={(e) => setSearchValue(e.target.value)}
-                        onPressEnter={handleSearch}
-                        allowClear
-                        onClear={() => setSearchValue('')}
-                    />
-
+                    <HeaderSearch />
                     <Menu
                         mode="inline"
                         items={navItems}
@@ -391,31 +223,6 @@ export default function HeaderCom({ categoryMenu = [], regionMenu = [] }: Header
                     />
                 </div>
             </Drawer>
-            {!screens.md && (
-                <Drawer
-                    title="Search"
-                    placement="top"
-                    onClose={toggleSearch}
-                    open={searchVisible}
-                    height="auto"
-                >
-                    <Input
-                        placeholder="Tìm kiếm"
-                        suffix={
-                            <SearchOutlined
-                                onClick={handleSearchIconClick}
-                                style={{ cursor: 'pointer' }}
-                            />
-                        }
-                        style={{ width: '100%' }}
-                        value={searchValue}
-                        onChange={(e) => setSearchValue(e.target.value)}
-                        onPressEnter={handleSearch}
-                        allowClear
-                        onClear={() => setSearchValue('')}
-                    />
-                </Drawer>
-            )}
         </Header>
     );
 }
